linear program works
okay so the linear program works. the all_constraints are the
constraints, the cost_obj_func is the objective function. the
linearProgram actually runs it. backend is just a script i used to run
the decisionsMichelle file (pousch it might be helpful for you). the
only thing left to be fixed is that the average cost from plant to
storage unit needs to be adjusted.